+++ src/client/src/commons/form/borehole/hydrogeology/hydrotestInput.js
@@ -19,6 +19,7 @@ import { useTranslation } from "react-i18next";
 import { useHydrotestDomains, useDomains } from "../../../../api/fetchApiV2";
 import { AlertContext } from "../../../alert/alertContext";
 import { ObservationType } from "./observationType";
+import { hydrogeologySchemaConstants } from "./hydrogeologySchemaConstants";
 
 const HydrotestInput = ({
   hydrotest,
@@ -190,7 +191,11 @@ const HydrotestInput = ({
                       }}
                       onMouseDown={handleMouseDown}>
                       {domains?.data
-                        ?.filter(d => d.schema === "htest101")
+                        ?.filter(
+                          d =>
+                            d.schema ===
+                            hydrogeologySchemaConstants.hydrotestKind,
+                        )
                         .sort((a, b) => a.order - b.order)
                         .map(d => (
                           <MenuItem key={d.id} value={d.id}>
@@ -209,7 +214,11 @@ const HydrotestInput = ({
                   control={control}
                   defaultValue={
                     hydrotest?.codelists
-                      ?.filter(c => c.schema === "htest102")
+                      ?.filter(
+                        c =>
+                          c.schema ===
+                          hydrogeologySchemaConstants.hydrotestFlowDirection,
+                      )
                       .map(c => c.id) || []
                   }
                   render={({ field }) => (
@@ -219,7 +228,9 @@ const HydrotestInput = ({
                       disabled={
                         !!formState.errors.testKindId ||
                         !filteredTestKindDomains?.data?.filter(
-                          d => d.schema === "htest102",
+                          d =>
+                            d.schema ===
+                            hydrogeologySchemaConstants.hydrotestFlowDirection,
                         ).length > 0
                       }
                       label={t("flowDirection")}
@@ -234,7 +245,11 @@ const HydrotestInput = ({
                         field.onChange(e.target.value);
                       }}>
                       {filteredTestKindDomains?.data
-                        ?.filter(d => d.schema === "htest102")
+                        ?.filter(
+                          d =>
+                            d.schema ===
+                            hydrogeologySchemaConstants.hydrotestFlowDirection,
+                        )
                         .map(d => (
                           <MenuItem key={d.id} value={d.id}>
                             {d[i18n.language]}
@@ -254,7 +269,11 @@ const HydrotestInput = ({
                   control={control}
                   defaultValue={
                     hydrotest?.codelists
-                      ?.filter(c => c.schema === "htest103")
+                      ?.filter(
+                        c =>
+                          c.schema ===
+                          hydrogeologySchemaConstants.hydrotestEvaluationMethod,
+                      )
                       .map(c => c.id) || []
                   }
                   render={({ field }) => (
@@ -264,7 +283,9 @@ const HydrotestInput = ({
                       disabled={
                         !!formState.errors.testKindId ||
                         !filteredTestKindDomains?.data?.filter(
-                          d => d.schema === "htest103",
+                          d =>
+                            d.schema ===
+                            hydrogeologySchemaConstants.hydrotestEvaluationMethod,
                         ).length > 0
                       }
                       label={t("evaluationMethod")}
@@ -279,7 +300,11 @@ const HydrotestInput = ({
                         field.onChange(e.target.value);
                       }}>
                       {filteredTestKindDomains?.data
-                        ?.filter(d => d.schema === "htest103")
+                        ?.filter(
+                          d =>
+                            d.schema ===
+                            hydrogeologySchemaConstants.hydrotestEvaluationMethod,
+                        )
                         .map(d => (
                           <MenuItem key={d.id} value={d.id}>
                             {d[i18n.language]}