+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added boreholes.swissgeol.ch id to details view.
 - Added `import_id` as required field for borehole import.
 - Added functionality to add, edit and delete hydrotests, which belong to the hydrogeology observations of a borehole.
+- Added functionality to add, edit and delete field measurement, which belong to the hydrogeology observations of a borehole.
 
 ### Changed
 
@@ -27,7 +28,7 @@
 ### Added
 
 - Defined borehole csv import fields `original_name`, `location_x` and `location_y` as required.
-- Added validation for required borehole csv import fields. 
+- Added validation for required borehole csv import fields.
 - Added online help.
 - Added validation for duplicated boreholes during the cvs import.
 - Added functionality to add, edit and delete wateringresses, which belong to the hydrogeology observations of a borehole.
@@ -80,7 +81,7 @@
 - Removed import settings from administrator settings.
 - Enabled editor mode settings for all user types.
 - Replaced csv import coordinates related columns by `location_x` and `location_y`.
-- Auto detect the provided coordinate reference system in the csv import and calculate the coordinates for the not provided reference system. 
+- Auto detect the provided coordinate reference system in the csv import and calculate the coordinates for the not provided reference system.
 
 ### Fixed